Rewrote to use a linked key list instead of a dqueue style rotating buffer.
The previous implementation didn't treat reads as uses for the
purposes of computing least recently used.

This fixes that.  Reads need to be lightweight and eviction free
so the rotating buffer wasn't viable.

Tested: Added an internal integrity check function and turned it on
for every step during test running along with an extended fuzzer
run.